examples/plugins/demo_plugin/requirements.txt
Pillow>=11,<12
